+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 v0.2.0 (in development)
 -----------------------
 - Support Python 3.9 and 3.10
+- Rename `EntryPoint.object` to `attr`.  The `object` name remains for
+  backwards compatibility but is deprecated.
 
 v0.1.0 (2020-07-20)
 -------------------